% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle.R
\name{smle}
\alias{smle}
\title{Performs efficient semiparametric estimation for regression models with univariate outcomes under general two-phase designs.}
\usage{
smle(Y = NULL, Delta = NULL, X = NULL, Z = NULL, Bspline_Z = NULL,
  data = NULL, hn_scale = 1, MAX_ITER = 2000, TOL = 1e-04,
  noSE = FALSE, model = "linear", verbose = FALSE)
}
\arguments{
\item{Y}{Specifies the column of the continuous, binary (\eqn{0} or \eqn{1}), or censored outcome. Subjects with missing values of \code{Y} are omitted from the analysis. This argument is required.}

\item{Delta}{Specifies the column of the event indicator. This argument is required when performing proportional hazards regression.}

\item{X}{Specifies the columns of the expensive covariates. Subjects with missing values of \code{X} are considered as those not selected in the second phase. This argument is required.}

\item{Z}{Specifies the columns of the inexpensive covariates. Subjects with missing values of \code{Z} are omitted from the analysis. This argument is optional.}

\item{Bspline_Z}{Specifies the columns of the B-spline basis. Subjects with missing values of \code{Bspline_Z} are omitted from the analysis. This argument is not needed when \code{X} is independent of \code{Z}.}

\item{data}{Specifies the name of the dataset. This argument is required.}

\item{hn_scale}{Specifies the scale of the perturbation constant in the variance estimation. For example, if \code{hn_scale = 0.5}, then the perturbation constant is \eqn{0.5n^{-1/2}}, where \eqn{n} is the first-phase sample size. The default value is \code{1}. This argument is optional. It is not needed when there is no \code{Z}.}

\item{MAX_ITER}{Specifies the maximum number of iterations in the EM algorithm. The default number is \code{2000}. This argument is optional.}

\item{TOL}{Specifies the convergence criterion in the EM algorithm. The default value is \code{1E-4}. This argument is optional.}

\item{noSE}{If \code{TRUE}, then the variances of the parameter estimators will not be estimated. The default value is \code{FALSE}. This argument is optional.}

\item{model}{Specifies the model. Possible values are "linear", "logistic", and "coxph". The default value is "linear".}

\item{verbose}{If \code{TRUE}, then show details of the analysis. The default value is \code{FALSE}.}
}
\value{
\item{coefficients}{Stores the analysis results.}
\item{converge}{In parameter estimation, if the EM algorithm converges, then \code{converge = TRUE}. Otherwise, \code{converge = FALSE}.}
\item{converge2}{In variance estimation, if the EM algorithm converges, then \code{converge2 = TRUE}. Otherwise, \code{converge2 = FALSE}.}
}
\description{
Performs efficient semiparametric estimation for regression models with univariate outcomes under general two-phase designs.
}
\references{
Tao, R., Zeng, D., and Lin, D. Y. (2017). "Efficient Semiparametric Inference Under Two-Phase Sampling, with Applications to Genetic Association Studies", Journal of the American Statistical Association, 112: 1468-1476.
}
