% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_MEXY.R
\name{smle_MEXY}
\alias{smle_MEXY}
\title{Performs efficient semiparametric estimation for general two-phase measurement error models when there are errors in both the outcome and covariates.}
\usage{
smle_MEXY(Y_tilde = NULL, Y = NULL, X_tilde = NULL, X = NULL,
  Z = NULL, data = NULL, MAX_ITER = 2000, TOL = 1e-04, noSE = FALSE,
  verbose = FALSE)
}
\arguments{
\item{Y_tilde}{Specifies the column of the error-prone outcome that is continuous. Subjects with missing values of \code{Y_tilde} are omitted from the analysis. This argument is required.}

\item{Y}{Specifies the column that stores the validated value of \code{Y_tilde} in the second phase. Subjects with missing values of \code{Y} are considered as those not selected in the second phase. This argument is required.}

\item{X_tilde}{Specifies the columns of the error-prone covariates. Subjects with missing values of \code{X_tilde} are omitted from the analysis. This argument is required.}

\item{X}{Specifies the columns that store the validated values of \code{X_tilde} in the second phase. Subjects with missing values of \code{X} are considered as those not selected in the second phase. This argument is required.}

\item{Z}{Specifies the columns of the accurately measured covariates. Subjects with missing values of \code{Z} are omitted from the analysis. This argument is optional.}

\item{data}{Specifies the name of the dataset. This argument is required.}

\item{MAX_ITER}{Specifies the maximum number of iterations in the EM algorithm. The default number is \code{2000}. This argument is optional.}

\item{TOL}{Specifies the convergence criterion in the EM algorithm. The default value is \code{1E-4}. This argument is optional.}

\item{noSE}{If \code{TRUE}, then the variances of the parameter estimators will not be estimated. The default value is \code{FALSE}. This argument is optional.}

\item{verbose}{If \code{TRUE}, then show details of the analysis. The default value is \code{FALSE}.}
}
\value{
\item{coefficients}{Stores the analysis results.}
\item{converge}{In parameter estimation, if the EM algorithm converges, then \code{converge = TRUE}. Otherwise, \code{converge = FALSE}.}
\item{converge2}{In variance estimation, if the EM algorithm converges, then \code{converge2 = TRUE}. Otherwise, \code{converge2 = FALSE}.}
}
\description{
Performs efficient semiparametric estimation for general two-phase measurement error models when there are errors in both the outcome and covariates.
}
